Move common code to launcher.go
