+++ private/app/launcher/launcher_windows.go
@@ -20,14 +20,11 @@ package launcher
 import (
 	"context"
 	"fmt"
-	"io"
-	"net"
 	"os"
 	"path/filepath"
 	"syscall"
 	"time"
 
-	"github.com/prometheus/client_golang/prometheus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"golang.org/x/sys/windows/svc"
@@ -35,11 +32,6 @@ import (
 	"golang.org/x/sys/windows/svc/eventlog"
 
 	"github.com/scionproto/scion/pkg/log"
-	"github.com/scionproto/scion/pkg/private/prom"
-	"github.com/scionproto/scion/pkg/private/serrors"
-	"github.com/scionproto/scion/private/app/command"
-	libconfig "github.com/scionproto/scion/private/config"
-	"github.com/scionproto/scion/private/env"
 )
 
 // Windows-specific configuration keys
@@ -56,44 +48,10 @@ const (
 
 // Application models a SCION server application.
 type Application struct {
-	// TOMLConfig holds the Go data structure for the application-specific
-	// TOML configuration. The Application launcher will check if the TOMLConfig
-	// supports additional methods (e.g., custom logging or instance ID) and
-	// extract them from the config if that is the case. See the XxxConfig interfaces
-	// in this package for more information.
-	TOMLConfig libconfig.Config
-
-	// Samplers contains additional configuration samplers to be included
-	// under the sample subcommand. If empty, no additional samplers are
-	// listed.
-	//
-	// DEPRECATED. This field will be removed once Anapaya/scion#5000 is implemented.
-	Samplers []func(command.Pather) *cobra.Command
-
-	// ShortName is the short name of the application. If empty, the executable name is used.
-	// The ShortName could be, for example, "SCION Daemon" for the SCION Daemon.
-	ShortName string
-
-	// RequiredIPs should return the IPs that this application wants to listen
-	// on. The launcher will wait until those IPs can be listened on. The
-	// function is called after the configuration has been  initialized. If this
-	// function is not set the launcher will immediately start the application
-	// without waiting for any IPs.
-	RequiredIPs func() ([]net.IP, error)
-
-	// Main is the custom logic of the application. If nil, no custom logic is executed
-	// (and only the setup/teardown harness runs). If Main returns an error, the
-	// Run method will return a non-zero exit code.
-	Main func(ctx context.Context) error
-
-	// ErrorWriter specifies where error output should be printed. If nil, os.Stderr is used.
-	ErrorWriter io.Writer
+	ApplicationBase
 
 	// cmd is the Cobra command for a SCION server application.
 	cmd *cobra.Command
-
-	// config contains the Viper configuration KV store.
-	config *viper.Viper
 
 	// isService indicates whether the application was started as a Windows service.
 	isService bool
@@ -239,45 +197,14 @@ loop:
 	return
 }
 
-func (a *Application) getShortName(executable string) string {
-	if a.ShortName != "" {
-		return a.ShortName
-	}
-	return executable
-}
-
 func (a *Application) executeCommand(ctx context.Context, shortName string) error {
-	os.Setenv("TZ", "UTC")
 
-	// Load launcher configurations from the same config file as the custom
-	// application configuration.
-	a.config.SetConfigType("toml")
-	a.config.SetConfigFile(a.config.GetString(cfgConfigFile))
-	if err := a.config.ReadInConfig(); err != nil {
-		return serrors.Wrap("loading generic server config from file", err,
-			"file", a.config.GetString(cfgConfigFile))
-
+	if err := a.ApplicationBase.loadConfig(); err != nil {
+		return err
 	}
-
-	if err := libconfig.LoadFile(a.config.GetString(cfgConfigFile), a.TOMLConfig); err != nil {
-		return serrors.Wrap("loading config from file", err,
-			"file", a.config.GetString(cfgConfigFile))
-
+	if err := a.ApplicationBase.initLogging(); err != nil {
+		return err
 	}
-	a.TOMLConfig.InitDefaults()
-
-	logEntriesTotal := prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "lib_log_emitted_entries_total",
-			Help: "Total number of log entries emitted.",
-		},
-		[]string{"level"},
-	)
-	opt := log.WithEntriesCounter(log.EntriesCounter{
-		Debug: logEntriesTotal.With(prometheus.Labels{"level": "debug"}),
-		Info:  logEntriesTotal.With(prometheus.Labels{"level": "info"}),
-		Error: logEntriesTotal.With(prometheus.Labels{"level": "error"}),
-	})
 
 	if a.config.GetString(cfgLogFile) != "" {
 		// Windows services do not have stdout and stderr, redirect to a log file instead.
@@ -291,48 +218,5 @@ func (a *Application) executeCommand(ctx context.Context, shortName string) erro
 		os.Stderr = logfile
 	}
 
-	if err := log.Setup(a.getLogging(), opt); err != nil {
-		return serrors.Wrap("initialize logging", err)
-	}
-	defer log.Flush()
-	if a.RequiredIPs != nil {
-		ips, err := a.RequiredIPs()
-		if err != nil {
-			return serrors.Wrap("loading required IPs", err)
-		}
-		WaitForNetworkReady(ctx, ips)
-	}
-	if err := env.LogAppStarted(shortName, a.config.GetString(cfgGeneralID)); err != nil {
-		return err
-	}
-	defer env.LogAppStopped(shortName, a.config.GetString(cfgGeneralID))
-	defer log.HandlePanic()
-
-	exportBuildInfo()
-	prom.ExportElementID(a.config.GetString(cfgGeneralID))
-	if err := a.TOMLConfig.Validate(); err != nil {
-		return serrors.Wrap("validate config", err)
-	}
-
-	if a.Main == nil {
-		return nil
-	}
-	return a.Main(ctx)
-}
-
-func (a *Application) getLogging() log.Config {
-	return log.Config{
-		Console: log.ConsoleConfig{
-			Level:           a.config.GetString(cfgLogConsoleLevel),
-			Format:          a.config.GetString(cfgLogConsoleFormat),
-			StacktraceLevel: a.config.GetString(cfgLogConsoleStacktraceLevel),
-		},
-	}
-}
-
-func (a *Application) getErrorWriter() io.Writer {
-	if a.ErrorWriter != nil {
-		return a.ErrorWriter
-	}
-	return os.Stderr
+	return a.ApplicationBase.executeCommand(ctx, shortName)
 }
